Rename commands -> command
