+++ command/src/main/java/discord4j/command/Command.java
@@ -1,4 +1,4 @@
-package discord4j.commands;
+package discord4j.command;
 
 import discord4j.core.event.domain.message.MessageCreateEvent;
 import reactor.core.publisher.Mono;
