+++ command/src/main/java/discord4j/command/DefaultCommandDispatcher.java
@@ -1,4 +1,4 @@
-package discord4j.commands;
+package discord4j.command;
 
 import discord4j.core.event.domain.message.MessageCreateEvent;
 import reactor.core.publisher.Flux;
@@ -7,10 +7,10 @@ import reactor.core.publisher.Mono;
 import java.util.*;
 
 /**
- * This is the default {@link discord4j.commands.CommandDispatcher} implementation used for
- * the {@link discord4j.commands.CommandBootstrapper}.
+ * This is the default {@link discord4j.command.CommandDispatcher} implementation used for
+ * the {@link discord4j.command.CommandBootstrapper}.
  *
- * @see discord4j.commands.CommandBootstrapper
+ * @see discord4j.command.CommandBootstrapper
  */
 public final class DefaultCommandDispatcher implements CommandDispatcher {
