+++ command/src/main/java/discord4j/command/package-info.java
@@ -1,9 +1,9 @@
 /**
  * This package contains a set of classes designed to allow for extreme customizability and interoperability when
- * building commands or command systems. While not essential for producing commands, it is recommended to conform to
+ * building command or command systems. While not essential for producing command, it is recommended to conform to
  * these apis in order to allow for interoperability between different implementations.
  */
 @NonNullApi
-package discord4j.commands;
+package discord4j.command;
 
 import reactor.util.annotation.NonNullApi;
