+++ command/src/test/java/discord4j/command/example/EchoCommand.java
@@ -1,6 +1,6 @@
-package discord4j.commands.example;
+package discord4j.command.example;
 
-import discord4j.commands.Command;
+import discord4j.command.Command;
 import discord4j.core.event.domain.message.MessageCreateEvent;
 import reactor.core.publisher.Mono;
