+++ command/src/test/java/discord4j/command/example/LogoutCommand.java
@@ -1,7 +1,7 @@
-package discord4j.commands.example;
+package discord4j.command.example;
 
-import discord4j.commands.Command;
-import discord4j.commands.CommandException;
+import discord4j.command.Command;
+import discord4j.command.CommandException;
 import discord4j.core.event.domain.message.MessageCreateEvent;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.util.Snowflake;
