+++ settings.gradle
@@ -3,4 +3,4 @@ include 'rest'
 include 'gateway'
 include 'common'
 include 'store'
-include 'commands'
+include 'command'
